% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-method-diagnose.R
\name{converged}
\alias{converged}
\title{Determine if a model converged.}
\usage{
converged(x, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A logical.
}
\description{
Determine if a model converged.
}
\examples{
my_model <- fit_model("vader", order = 66)
converged(my_model)
}
