% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mod.R
\name{fit_model}
\alias{fit_model}
\title{Fit a fake model}
\usage{
fit_model(name, ...)
}
\arguments{
\item{name}{The name of the model.}

\item{...}{Additional arguments.}
}
\value{
A `pkga_fit` object.
}
\description{
Fit a fake model
}
\examples{
fit_model("taylor", lucky = 13, year = 1989)
}
